+++ app/services/merge_requests/base_service.rb
@@ -94,6 +94,10 @@ module MergeRequests
 
     private
 
+    def refresh_pipelines_on_merge_requests(merge_request, allow_duplicate: false)
+      create_pipeline_for(merge_request, current_user, async: true, allow_duplicate: allow_duplicate)
+    end
+
     def enqueue_jira_connect_messages_for(merge_request)
       return unless project.jira_subscription_exists?