+++ app/services/merge_requests/refresh_service.rb
@@ -162,10 +162,6 @@ module MergeRequests
       @outdate_service ||= Suggestions::OutdateService.new
     end
 
-    def refresh_pipelines_on_merge_requests(merge_request)
-      create_pipeline_for(merge_request, current_user, async: true)
-    end
-
     def abort_auto_merges(merge_request)
       abort_auto_merge(merge_request, 'source branch was updated')
     end