+++ spec/frontend/boards/components/board_content_spec.js
@@ -60,7 +60,6 @@ describe('BoardContent', () => {
     wrapper = shallowMount(BoardContent, {
       apolloProvider: fakeApollo,
       propsData: {
-        disabled: false,
         boardId: 'gid://gitlab/Board/1',
         ...props,
       },
@@ -72,6 +71,7 @@ describe('BoardContent', () => {
         isIssueBoard,
         isEpicBoard,
         isGroupBoard: true,
+        disabled: false,
         isApolloBoard,
       },
       store,